Quick fix: update GH actions to repeat build step
The proper fix for this is to shift Docker to volumes and run our build
step on the server directly after running it in GH actions (for
redundancy). For now, I'm hoping just running it in the same job as the
Docker publish will give it access to the files there.