Add match start time with automatic updating for rundown
Fair warning, this commit is throwing a bunch of typescript errors for
me locally even though there are no actual errors in the built code. I
cannot for the life of me figure out where the root error lies, given
that even a file that did not change was throwing them at first.